+++ py_experimenter/experimenter.py
@@ -12,10 +12,7 @@ from py_experimenter import utils
 from py_experimenter.config import PyExperimenterCfg
 from py_experimenter.database_connector_lite import DatabaseConnectorLITE
 from py_experimenter.database_connector_mysql import DatabaseConnectorMYSQL
-from py_experimenter.exceptions import (
-    InvalidConfigError,
-    NoExperimentsLeftException,
-)
+from py_experimenter.exceptions import InvalidConfigError, NoExperimentsLeftException
 from py_experimenter.experiment_status import ExperimentStatus
 from py_experimenter.result_processor import ResultProcessor
 
@@ -29,7 +26,7 @@ class PyExperimenter:
         self,
         experiment_configuration_file_path: str = os.path.join("config", "experiment_configuration.yml"),
         database_credential_file_path: str = os.path.join("config", "database_credentials.yml"),
-        use_ssh_tunnel: bool = False,
+        use_ssh_tunnel: Optional[bool] = None,
         table_name: str = None,
         database_name: str = None,
         use_codecarbon: bool = True,
@@ -48,10 +45,12 @@ class PyExperimenter:
         :param database_credential_file_path: The path to the database configuration file storing the credentials
             for the database connection, i.e., host, user and password. Defaults to 'config/database_credentials.cfg'.
         :type database_credential_file_path: str, optional
-        :param use_ssh_tunnel: If the used dataabse is sqlite this parameter is ignored Otherwise: If the database is mysql,
-            and `use_ssh_tunnel == True` the ssh credentials provided in `database_credential_file_path` used to establish
-            a ssh tunnel to the database. If `use_ssh_tunnel == True` but no ssh credentials are provided in
-            `database_credential_file_path`, no ssh tunnel is established. Defaults to True.
+        :param use_ssh_tunnel: If the used database is sqlite this parameter is ignored. Otherwise: If the database is mysql,
+            and `use_ssh_tunnel == None` the ssh decision is based on the configuration file (defaults to false).
+            If `use_ssh_tunnel != True` the ssh credentials provided in `database_credential_file_path` are used to establish
+            an ssh tunnel to the database. If `use_ssh_tunnel == True` but no ssh credentials are provided in
+            `database_credential_file_path`, an error is raised. If `use_shh_tunnel==False` PxExperimenter directly connects
+            to the databse. Defaults to None.
         :type use_ssh_tunnel: bool
         :param table_name: The name of the database table, if given it will overwrite the table_name given in the
             `experiment_configuration_file_path`. If None, the table table name is taken from the experiment
@@ -74,6 +73,7 @@ class PyExperimenter:
         :type log_file: str
         :raises InvalidConfigError: If either the experiment or database configuration are missing mandatory information.
         :raises ValueError: If an unsupported or unknown database connection provider is given.
+        :raises SshTunnelError: If the ssh tunnel could not be established, or if the ssh credentials are missing/invalid.
         """
         # If the logger is not allready craeted, create it with the given name and level
         self.logger_name = logger_name
@@ -104,7 +104,10 @@ class PyExperimenter:
             raise InvalidConfigError("Invalid configuration")
 
         self.database_credential_file_path = database_credential_file_path
-        self.use_ssh_tunnel = use_ssh_tunnel
+
+        # If use_ssh_tunnel is not None, the decision is based on the given kwarg
+        if use_ssh_tunnel is not None:
+            self.config.database_configuration.use_ssh_tunnel = use_ssh_tunnel
 
         if table_name is not None:
             self.config.database_configuration.table_name = table_name
@@ -118,7 +121,7 @@ class PyExperimenter:
             self.db_connector = DatabaseConnectorLITE(self.config.database_configuration, self.use_codecarbon, self.logger)
         elif self.config.database_configuration.provider == "mysql":
             self.db_connector = DatabaseConnectorMYSQL(
-                self.config.database_configuration, self.use_codecarbon, database_credential_file_path, use_ssh_tunnel, self.logger
+                self.config.database_configuration, self.use_codecarbon, database_credential_file_path, self.logger
             )
         else:
             raise ValueError("The provider indicated in the config file is not supported")
@@ -129,7 +132,7 @@ class PyExperimenter:
         """
         Closes the ssh tunnel if it is used.
         """
-        if self.config.database_configuration.provider == "mysql" and self.use_ssh_tunnel:
+        if self.config.database_configuration.provider == "mysql":
             self.db_connector.close_ssh_tunnel()
         else:
             self.logger.warning("No ssh tunnel to close")