+++ CHANGELOG.rst
@@ -10,6 +10,13 @@ Feature
 -------
 
 - Added documentation about how to execute PyExperimenter on distributed machines.
+- Improved the usage and documentation of ssh tunnel to be more flexible and user friendly.
+
+Fix
+---
+
+
+
 
 v1.4.1 (11.03.2024)
 ===================
@@ -20,7 +27,6 @@ Feature
 - Added logging whenever a sql statement is executed.
 - Update CI such that the tests are run on Python 3.9 and 3.10, additionally removed the paper draft action. 
 - Added logging when initializing database to contain resultfields and logfields.
-- Adapt ssh tunnel usage to be more flexible and user friendly.
 
 Fix
 ---