+++ CHANGELOG.rst
@@ -1,7 +1,13 @@
 =========
 Changelog
 =========
+v1.4.3 (**.**.2025)
+===================
 
+Fix
+===
+
+- Fix bug, where the logtable_name was not overwritten by `table_name` updates in the `PyExperimenter` class.
 
 v1.4.2 (12.06.2024)
 ===================